+++ apps/server/src/app/app.module.ts
@@ -105,11 +105,7 @@ const BOT_NAMES_RU = env.get('BOT_NAMES_RU').required().asArray();
       },
       botMeetingInformation: {
         en: [`Hello! I'm ${BOT_NAMES[0]} 😉`, 'Hello!', 'Hello 🖖'],
-        ru: [
-          `Всем привет! я ${BOT_NAMES_RU[0]} 😉`,
-          `Всем привет!`,
-          'Всем привет 🖖',
-        ],
+        ru: [`Привет! я ${BOT_NAMES_RU[0]} 😉`, `Привет!`, 'Привет 🖖'],
       },
     }),
     LanguageSwitherModule.forRoot(),