+++ libs/schematics/src/generators/init/lib/index.ts
@@ -1,0 +1,3 @@
+export * from './add-custom';
+export * from './add-dependencies';
+export * from './normalize-options';