+++ libs/schematics/src/public.ts
@@ -1,0 +1,3 @@
+export { applicationGenerator } from './generators/application/application';
+export { initGenerator } from './generators/init/init';
+export { libraryGenerator } from './generators/library/library';